Split the tab stops for the `struct` snippet to be more like `fn`
I think it's nice to have the tabs correspond to each individual field
only, so that you think about the name and the type separately, and that
you don't have to remember where the colon goes if you use the first tab
field that was here that replaces everything.

Also have `type` instead of defaulting to `int` (which doesn't even
exist anymore in rust).