Improve cargo commands
Makes the cargo commands produce output in an HTML window, so that error messages are clickable. It also adds cargo check, cargo fmt and cargo clippy commands.

To support so many new commands the bundle now uses the single shortcut key approach.

Also has changes under the hood to support using the bundle for Fuchsia development and a refinement of the symbol list.